LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-single-value-in-shadow-tree-expected.txt


FAIL Meta color-scheme in shadow tree does not apply. assert_equals: expected (string) "normal" but got (undefined) undefined
